only sort cards once
